libz_rs_sys

Function compress2

pub unsafe extern "C-unwind" fn compress2(
    dest: *mut Bytef,
    destLen: *mut c_ulong,
    source: *const Bytef,
    sourceLen: c_ulong,
    level: c_int,
) -> c_int

Compresses source into dest, and writes the final deflated size into destLen.

The level parameter has the same meaning as in deflateInit_. sourceLen is the byte length of the source buffer. Upon entry, destLen is the total size of the destination buffer, which must be at least the value returned by compressBound(sourceLen). Upon exit, destLen is the actual size of the compressed data.

Returns

Z_OK if success

Z_MEM_ERROR if there was not enough memory

Z_BUF_ERROR if there was not enough room in the output buffer

Safety

The caller must guarantee that

Either

destLen is NULL

destLen satisfies the requirements of &mut *destLen

Either

dest is NULL

dest and *destLen satisfy the requirements of core::slice::from_raw_parts_mut::<MaybeUninit<u8>>

Either

source is NULL

source and sourceLen satisfy the requirements of core::slice::from_raw_parts