libz_rs_sys

Function inflateBack

pub unsafe extern "C-unwind" fn inflateBack(
    _strm: z_streamp,
    _in: in_func,
    _in_desc: *mut c_void,
    _out: out_func,
    _out_desc: *mut c_void,
) -> c_int

Decompresses as much data as possible, and stops when the input buffer becomes empty or the output buffer becomes full.

Safety

The caller must guarantee that

Either

strm is NULL

strm satisfies the requirements of &mut *strm and was initialized with inflateBackInit_