libz_rs_sys

Function inflateBackEnd

pub unsafe extern "C-unwind" fn inflateBackEnd(
    _strm: z_streamp,
) -> c_int

Deallocates all dynamically allocated data structures for this stream.

This function discards any unprocessed input and does not flush any pending output.

Returns

Z_OK if success

Z_STREAM_ERROR if the stream state was inconsistent

Safety

The caller must guarantee that

Either

strm is NULL

strm satisfies the requirements of &mut *strm and was initialized with inflateBackInit_